% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ordinals.R
\name{as.ordinal}
\alias{as.ordinal}
\title{Converts a number to an ordinal (e.g., first, second, etc.)}
\usage{
as.ordinal(num)
}
\arguments{
\item{number.}{The number to convert to ordinal.}
}
\description{
Converts a number to an ordinal (e.g., first, second, etc.)
}

