% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/file-cabinet.R
\name{create_file}
\alias{create_file}
\title{Creates a file from a value, with overwrite protection.}
\usage{
create_file(name, value, can_overwrite = FALSE)
}
\arguments{
\item{name}{character. The name of the file to create.}

\item{value}{ANY. The value to put in the file.}

\item{can_overwrite}{logical. If true, will replace an existing directory.}
}
\description{
Creates a file from a value, with overwrite protection.
}

