% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/file-cabinet.R
\name{create_folder}
\alias{create_folder}
\title{Creates a folder, with overwrite protection.}
\usage{
create_folder(dir, can_overwrite = FALSE)
}
\arguments{
\item{dir}{character. The name of the directory to create.}

\item{can_overwrite}{logical. If true, will replace an existing directory.}
}
\description{
Creates a folder, with overwrite protection.
}

