% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fn-utils.R
\name{delay_by}
\alias{delay_by}
\title{Delays the processing of a function by a certain amount of seconds.}
\usage{
delay_by(f, delay)
}
\arguments{
\item{f}{function. The function to run, with a delay.}

\item{delay}{numeric. The number of seconds to delay.}
}
\value{
a modified version of f to include a delay prior to running.
}
\description{
Useful for handling API calls and other processes where you don't want to overload things.
Taken from http://adv-r.had.co.nz/Function-operators.html by Hadley Wickham.
Intended to be used with magrittr <https://github.com/smbache/magrittr>
}
\examples{
2 \%>\% (identity \%>\% delay_by(10))
}

