% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interactive.R
\name{ensure_in_interactive}
\alias{ensure_in_interactive}
\title{Stop if the user is not in interactive mode.
Prevents code from being run outside the R console.}
\usage{
ensure_in_interactive()
}
\description{
Stop if the user is not in interactive mode.
Prevents code from being run outside the R console.
}

