% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interactive.R
\name{ensure_not_in_interactive}
\alias{ensure_not_in_interactive}
\title{Stop if the user is in interactive mode.
Prevents code from being run inside the R console.}
\usage{
ensure_not_in_interactive()
}
\description{
Stop if the user is in interactive mode.
Prevents code from being run inside the R console.
}

