% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/open_github.R
\name{get_github}
\alias{get_github}
\title{Finds the URL for a GitHub page.}
\usage{
get_github(package, fun = NULL, author = "cran", branch = "master",
  cap_r = TRUE)
}
\arguments{
\item{package}{character. The name of the R package (e.g., 'batbelt').
Can also be 'author/package' (e.g., 'peterhurford/batbelt').}

\item{fun}{character. The name of the function to open.  Leave NULL to just get the URL for the package overall.}

\item{author}{character. The GitHub username of the package.  Defaults to finding packages via cran.}

\item{branch}{character. The branch to look for.}

\item{cap_r}{logical. TRUE if the file ends with a capital R, FALSE otherwise.}
}
\description{
Finds the URL for a GitHub page.
}

