% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/open_github.R
\name{github_fun_search}
\alias{github_fun_search}
\title{Searches for a function in GitHub using GitHub's search.}
\usage{
github_fun_search(package, fun, author = "cran")
}
\arguments{
\item{package}{character. The name of the R package (e.g., 'batbelt').
Can also be 'author/package' (e.g., 'peterhurford/batbelt').}

\item{fun}{character. The name of the function to open.  Leave NULL to just get the URL for the package overall.}

\item{author}{character. The GitHub username of the package.  Defaults to finding packages via cran.}
}
\description{
Searches for a function in GitHub using GitHub's search.
}

