% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fn-utils.R
\name{progress_bar}
\alias{progress_bar}
\title{Displays a progress bar for a given function.}
\usage{
progress_bar(f, bar = "default", default.n = 1)
}
\arguments{
\item{f}{function. The function to run with the progress bar.}

\item{bar}{character. The name of the progress bar to use.  Default ('default') is to display a "." every 1 function run.}

\item{default.n}{integer. If using the default progress bar, display every n function runs instead.}
}
\description{
Taken from http://adv-r.had.co.nz/Function-operators.html by Hadley Wickham.
Intended to be used with magrittr <https://github.com/smbache/magrittr>
}
\examples{
seq(10) \%>\% (identity \%>\% progress_bar('default', 2))
}

